Fix MSVC warnings in ssl-related src

+++ src/mongoc/mongoc-stream-tls.c
@@ -468,7 +468,7 @@ _mongoc_stream_tls_write (mongoc_stream_tls_t *tls,
       expire = bson_get_monotonic_time () + (tls->timeout_msec * 1000UL);
    }
 
-   ret = BIO_write (tls->bio, buf, buf_len);
+   ret = BIO_write (tls->bio, buf, (int)buf_len);
 
    if (ret < 0) {
       return ret;
@@ -478,7 +478,7 @@ _mongoc_stream_tls_write (mongoc_stream_tls_t *tls,
       now = bson_get_monotonic_time ();
 
       if ((expire - now) < 0) {
-         if (ret < buf_len) {
+         if (ret < (ssize_t)buf_len) {
             mongoc_counter_streams_timeout_inc();
 #ifdef _WIN32
             errno = WSAETIMEDOUT;
@@ -489,7 +489,7 @@ _mongoc_stream_tls_write (mongoc_stream_tls_t *tls,
 
          tls->timeout_msec = 0;
       } else {
-         tls->timeout_msec = (expire - now) / 1000L;
+         tls->timeout_msec = (int32_t)(expire - now) / 1000L;
       }
    }
 
@@ -559,7 +559,7 @@ _mongoc_stream_tls_writev (mongoc_stream_t *stream,
       while (iov_pos < iov[i].iov_len) {
          if (buf_head != buf_tail ||
              ((i + 1 < iovcnt) &&
-              ((buf_end - buf_tail) > (iov[i].iov_len - iov_pos)))) {
+              ((size_t)(buf_end - buf_tail) > (iov[i].iov_len - iov_pos)))) {
             /* If we have either of:
              *   - buffered bytes already
              *   - another iovec to send after this one and we don't have more
@@ -567,7 +567,7 @@ _mongoc_stream_tls_writev (mongoc_stream_t *stream,
              *
              * copy into the buffer */
 
-            bytes = BSON_MIN (iov[i].iov_len - iov_pos, buf_end - buf_tail);
+            bytes = BSON_MIN (iov[i].iov_len - iov_pos, (size_t)(buf_end - buf_tail));
 
             memcpy (buf_tail, (char *) iov[i].iov_base + iov_pos, bytes);
             buf_tail += bytes;
@@ -603,7 +603,7 @@ _mongoc_stream_tls_writev (mongoc_stream_t *stream,
 
             ret += child_ret;
 
-            if (child_ret < to_write_len) {
+            if (child_ret < (ssize_t)to_write_len) {
                /* we timed out, so send back what we could send */
 
                return ret;
@@ -704,7 +704,7 @@ _mongoc_stream_tls_readv (mongoc_stream_t *stream,
 
                tls->timeout_msec = 0;
             } else {
-               tls->timeout_msec = (expire - now) / 1000L;
+               tls->timeout_msec = (int32_t)(expire - now) / 1000L;
             }
          }
